board repo check

+++ Mandatum/Application/EFRepository/BoardRepo.cs
@@ -16,53 +16,32 @@ namespace Application
 
         public void SaveBoard(BoardRecord board)
         {
-            try
-            {
-                _dbContext.Boards.Add(board);
-                _dbContext.SaveChanges();
-            }
-            catch
-            {
-                //
-            }
+            _dbContext.Boards.Add(board);
+            _dbContext.SaveChanges();
         }
 
         public BoardRecord GetBoard(Guid boardId)
         {
-            try
-            {
-                return _dbContext.Boards.AsNoTracking().Include(prop => prop.TaskIds).FirstOrDefault(board => board.Id == boardId);
-            }
-            catch
-            {
-                return new BoardRecord();
-            }
+            var boardRecord = _dbContext.Boards.AsNoTracking().Include(prop => prop.TaskIds)
+                .FirstOrDefault(board => board.Id == boardId);
+            if (boardRecord is null) return new BoardRecord();
+            return boardRecord;
         }
 
         public void UpdateBoard(BoardRecord updBoard)
         {
-            try
-            {
-                _dbContext.Boards.Update(updBoard);
-                _dbContext.SaveChanges();
-            }
-            catch
-            {
-                //
-            }
+            var containsBoard = GetBoard(updBoard.Id);
+            if (containsBoard is null) return;
+            _dbContext.Boards.Update(updBoard);
+            _dbContext.SaveChanges();
         }
 
         public void DeleteBoard(BoardRecord board)
         {
-            try
-            {
-                _dbContext.Boards.Remove(board);
-                _dbContext.SaveChanges();
-            }
-            catch
-            {
-                //
-            }
+            var containsBoard = GetBoard(board.Id);
+            if (containsBoard is null) return;
+            _dbContext.Boards.Remove(board);
+            _dbContext.SaveChanges();
         }
     }
 }
